주석 저장

+++ webserver/dog/models.py
@@ -3,13 +3,18 @@ import re
 
 from django.db import models
 
-# Create your models here.
+# 개의 클래스.
 class Dog(models.Model):
-    dog_id = models.IntegerField(primary_key=True)
-    dog_picture_counter = models.PositiveIntegerField(default = 0)
+    # Contract에서 배정된 개 id
+    dog_id = models.PositiveIntegerField(primary_key = True)
+    # 개 등록 사진의 일련번호
+    dog_picture_counter = models.PositiveSmallIntegerField(default = 0)
+    # 개의 털길이(Enum)
     dog_coat_length = models.PositiveSmallIntegerField(null=True, blank=True)
+    # 개의 털색깔(String)
     dog_coat_color = models.CharField(max_length=100, null=True, blank=True)
 
+# 개의 사진에 대한 이름 정책. (개 id/사진 일련번호.확장자)
 def pic_name_policy(instance, filename):
     instance.dog.dog_picture_counter += 1
     return '%s/%s.%s' % (
@@ -18,10 +23,12 @@ def pic_name_policy(instance, filename):
         filename.split('.')[-1]
         )
 
+# 개 사진의 url 테이블.
 class Picture(models.Model):
     dog = models.ForeignKey(Dog, on_delete=models.CASCADE)
     picture_url = models.ImageField(upload_to=pic_name_policy)
 
+# 개 종류의 테이블, 운영자만 수정 가능.
 class Breed(models.Model):
     breed_kind = models.CharField(max_length=100)
     breed_size = models.PositiveSmallIntegerField()
